docker/mqtt2http/requirements.txt
paho-mqtt==1.6.*
cbor2==5.*
